DOTOLI-281 eslint-config base·next 설정 분리
