+++ .changeset/rn7ozl07.md
@@ -1,0 +1,5 @@
+---
+'@bbodek/eslint-config': patch
+---
+
+DOTOLI-280 eslint-config react-hooks·jsx-a11y 룰 추가
