+++ .changeset/vf4zhlht.md
@@ -1,0 +1,5 @@
+---
+'@bbodek/eslint-config': patch
+---
+
+DOTOLI-281 eslint-config base·next 설정 분리
